Pull all platform specific code into FastFetch's directory. Refactor PrefetchLibGit2 and move it into FastFetch. Rename PrefetchLibGit2Repo -> FastFetchLibGit2Repo
