+++ GVFS/GVFS.Platform.Mac/MacFileSystem.cs
@@ -58,40 +58,11 @@ namespace GVFS.Platform.Mac
             return NativeStat.IsSock(statBuffer.Mode);
         }
 
-        public unsafe void WriteFile(ITracer tracer, byte* originalData, long originalSize, string destination, ushort mode)
-        {
-            int fileDescriptor = -1;
-            try
-            {
-                fileDescriptor = NativeFileReader.Open(destination, NativeFileReader.WriteOnly | NativeFileReader.Create, mode);
-                if (fileDescriptor != -1)
-                {
-                    IntPtr result = Write(fileDescriptor, originalData, (IntPtr)originalSize);
-                    if (result.ToInt32() == -1)
-                    {
-                        throw new Win32Exception(Marshal.GetLastWin32Error());
-                    }
-                }
-            }
-            catch (Win32Exception e)
-            {
-                tracer.RelatedError("Error writing file {0}. Win32Exception: {1}", destination, e);
-                throw;
-            }
-            finally
-            {
-                NativeFileReader.Close(fileDescriptor);
-            }
-        }
-
         [DllImport("libc", EntryPoint = "chmod", SetLastError = true)]
         private static extern int Chmod(string pathname, ushort mode);
 
         [DllImport("libc", EntryPoint = "rename", SetLastError = true)]
         private static extern int Rename(string oldPath, string newPath);
-
-        [DllImport("libc", EntryPoint = "write", SetLastError = true)]
-        private static unsafe extern IntPtr Write(int fileDescriptor, void* buf, IntPtr count);
 
         private NativeStat.StatBuffer StatFile(string fileName)
         {
@@ -199,9 +170,6 @@ namespace GVFS.Platform.Mac
 
             [DllImport("libc", EntryPoint = "open", SetLastError = true)]
             public static extern int Open(string path, int flag);
-
-            [DllImport("libc", EntryPoint = "open", SetLastError = true)]
-            public static extern int Open(string path, int flag, int creationMode);
 
             [DllImport("libc", EntryPoint = "close", SetLastError = true)]
             public static extern int Close(int fd);
